requirements_py310.txt
ray==1.13.0
numpy==1.25.2
gym==0.17.1
gymnasium==0.29.1
atari-py==0.2.6
cython
torch==2.3.1
PyOpenGL-accelerate
hydra-core
protobuf==3.20.1
wandb==0.15.2
dm_control

opencv-python==4.8.0.74
kornia==0.6.1
imageio[ffmpeg]
colorednoise
line_profiler
setuptools==69.5.1
packaging==25.0
pydantic==1.10.13